refactor: とりあえずコンポーネントに分割しきった
