+++ src/components/Organisims/NewTermSetDialog.tsx
@@ -1,0 +1,89 @@
+"use client";
+
+import { Dialog, Flex, Text, TextField, Button } from "@radix-ui/themes";
+import { useState } from "react";
+import { createTermSet } from "@/repositories/termSetService";
+import { useAuth } from "@/contexts/AuthContext";
+
+type NewTermSetDialogProps = {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  onTermSetCreated: () => void;
+};
+
+export default function NewTermSetDialog({
+  open,
+  onOpenChange,
+  onTermSetCreated,
+}: NewTermSetDialogProps) {
+  const { user } = useAuth();
+  const [newTermTitle, setNewTermTitle] = useState("");
+  const [newTermDescription, setNewTermDescription] = useState("");
+  const [isCreating, setIsCreating] = useState(false);
+
+  const handleCreateTerm = async () => {
+    if (!user || !newTermTitle.trim()) return;
+
+    setIsCreating(true);
+    try {
+      await createTermSet(user.uid, newTermTitle.trim(), newTermDescription.trim());
+      onTermSetCreated();
+      setNewTermTitle("");
+      setNewTermDescription("");
+      onOpenChange(false);
+    } catch (error) {
+      console.error("利用規約の作成に失敗しました:", error);
+    } finally {
+      setIsCreating(false);
+    }
+  };
+
+  return (
+    <Dialog.Root open={open} onOpenChange={onOpenChange}>
+      <Dialog.Content style={{ maxWidth: 450 }}>
+        <Dialog.Title>新しい利用規約を作成</Dialog.Title>
+        <Dialog.Description size="2" mb="4">
+          利用規約のタイトルと説明を入力してください。
+        </Dialog.Description>
+
+        <Flex direction="column" gap="3">
+          <label>
+            <Text as="div" size="2" mb="1" weight="bold">
+              タイトル
+            </Text>
+            <TextField.Root
+              placeholder="例: サークル会則"
+              value={newTermTitle}
+              onChange={e => setNewTermTitle(e.target.value)}
+            />
+          </label>
+          <label>
+            <Text as="div" size="2" mb="1" weight="bold">
+              説明（任意）
+            </Text>
+            <TextField.Root
+              placeholder="例: サークル活動に関する規約"
+              value={newTermDescription}
+              onChange={e => setNewTermDescription(e.target.value)}
+            />
+          </label>
+        </Flex>
+
+        <Flex gap="3" mt="4" justify="end">
+          <Dialog.Close>
+            <Button variant="soft" color="gray">
+              キャンセル
+            </Button>
+          </Dialog.Close>
+          <Button
+            onClick={handleCreateTerm}
+            disabled={!newTermTitle.trim() || isCreating}
+            className="text-white"
+          >
+            {isCreating ? "作成中..." : "作成"}
+          </Button>
+        </Flex>
+      </Dialog.Content>
+    </Dialog.Root>
+  );
+}
